kernel - Use intmax_t when printing memory amounts
* Now that vm_pindex_t is 64 bits, fix various printf()s

+++ sys/platform/pc64/x86_64/machdep.c
@@ -367,9 +367,9 @@ again:
 	cninit();		/* the preferred console may have changed */
 #endif
 
-	kprintf("avail memory = %lu (%lu MB)\n",
-		ptoa(vmstats.v_free_count),
-		ptoa(vmstats.v_free_count) / 1024 / 1024);
+	kprintf("avail memory = %ju (%ju MB)\n",
+		(uintmax_t)ptoa(vmstats.v_free_count),
+		(uintmax_t)ptoa(vmstats.v_free_count) / 1024 / 1024);
 
 	/*
 	 * Set up buffers, so they can be used to read disk labels.
